define missing path sets
